Rename the id() method on WiegandCard to card_id().
Problem: For #11 completion we need to be able to map
credentials directly to schedules.
This means that credentials needs to be uniquely
identified: an id field make sense.

Solution: Rename the current method to get the card id to
card_id() so we can add a general id() method later.