Introduce new way to declare credentials. Not backward-compatible.
See the documentation update for the new format.
It is more flexible and will allow the implementation of
restricted credentials (with validity limit) and the direct
mapping of credcentials to schedule.

This commits updates the configuration tests file to reflect
this new format.
The implementation partiually support this new format.