Add a AUTODETECT mode for Wiegand Module.
When is mode is selected, the module will attempt to autodetect
the output mode of the reader and will support reading
different types of credentials.

This is a good mode if you don't know how your reader is configured
or if you want some flexibility in what it can process.