Introduce CoreAPI, a safe wrapper around Core API.
CoreAPI is a new object that allows retrieval of information
from the Core in a safe way. It uses the tasks infrastructure
internaly to run code in the main thread, and extract
required information.

This commit also improve the shutdown process. It's possible
for shutdown to trigger a deadlock:
    + If the core waits on a module, and the module waits on a
    task that run on the main thread, we get a deadlock.
    + Similarly, if the module sends messages to the core
    while the core waits for the module to shutdown, we get
    a deadlock.

The improved Leosac::Kernel code will now *request* shutdown,
and then continue normal operations for 5 seconds before
waiting on module's shutdown. It's not perfect, but its a
step in the right direction.