Rework user/group membership.
Instead of direct bidirectional relationship between a user and
a group, we now use a UserGroupMembership to tie them together.

This UserGroupMembership has the advantage of being able to hold
additional information about the membership, such as the rank
of the user inside the group, or the date at which the user
joined the group.