Introduces extensible Audit log.
While this is a work in progress, the audit log should be
extensible through modules. This is a first step towards that
goal.

Modules should be able to extends the Audit::AuditEntry object
and ODB will take care of the rest.
The Audit::Serializer infrastructure has been improved to support
runtime registered serializer: serializers that are provide by modules.

The Audit::Serializer::PolymorphicAuditJSON now has a
`register_serializer()` static method for use by modules.