Refactor AuditEntry serializers.
Leosac gains a new class, the ServiceRegistry. The goal is to allows
dynamic registration of various services. The ServiceRegistry
is generic enough and doesn't know nor care about the kind of
services it stores.

The runtime-registered audits serializer now use this new
infrastructure. Audit::Serializer::JSONService is the service providing
support for dynamically registered serializer.