Infrastructure change WRT AccessPoint WebSock handlers (P1)
The websocket API calls regarding AccessPoint is improved.
This commit introduces an AccessPointBackend interface object.
A new AccessPointService is being introduce, which allows
for runtime registered serializers, aswel as registration
of AccessPointBackend.

AccessPoint modules are expected to extend this object and
provide an implementation for the CRUD operations related
to their extended AccessPoint object.