Problem: Door and AccessPoint relationship.
Currently it was possible for multiple doors to
be linked to the same AccessPoint. Somehow this
caused issue when retrieving access points objects,
probably because the inverse relationship (AP to door)
was invalid.

The solution is to ensure, at the database level,
that doors cannot reference the same access point.