[CORE/DB] Fix std::time_point storage in SQLite.
Problem: Loading from database a timepoint
that was saved when it was equal to timepoint::max()
or timepoint::min() can cause "time overflow".

Because of timezone issue, loading a string version of
timepoint::max (with timezone GMT) on a system whose
timezone is GMT+2 doesn't work as we want.
We would load timepoint::max() then adds 2 hours, and
end up with something close to timepoint::min().

Solution: Literally store "TIME_POINT_MAX" and
"TIME_POINT_MIN" so we can handle this special case
and assign timepoint::max() or timepoint::min()
to the loaded timepoint.

This fixes issue #85.