Add noise filtering in authentication
Sometimes kickback will occur due to the nature of
inductors. Solenoid strikers send kicback through
the lines when they close. Additionally, wiegand
lines can be very sensitive to noise. As a result,
noise can come into the D0 and D1 lines and result
in false negatives.

This commit adds config variables to allow the user
to ignore a certain amount of bits coming in that
they might consider noise to have the software
ignore it as an access attempt.