Prevent always closed from denying access events
Before - always_closed schedules would prevent access to
the door even if the user was authorized. I've removed this
because this should be configured instead. If the door
shouldn't have access to it then this should be configured
in the schedule mapping itself for the user(s) that should
not be allowed access during that time.

Ultimately, there likely is never going to be a door
that should be denied access by all users ever. And if
there is, then it can be configured so in the schedule
mapping to prevent user access during a certain timeframe.