Auth DB Module (#148)
* Begin auth-db module

* Set up constructor for AuthDBInstance

* Fix log message

* Update cmake list

* Fix namespace typo

* Add missing virtual function

* Implement on_service_event

* Finish AuthDBModule declarations

* Add config setup for AuthDB

* Fix log for config setup

* Refactor

* Add AuthResult struct

* Start handle_bus_msg implementation

* Implement update_and_log_auth_result_msg

* Implement handle_kernel_msg

* Begin handle_auth implementation

* Update handle_auth to pass by pointer

* Set up fetch credentials from database structure

Checks credentials passed in message, and if they are
RFIDCard credentials, check to see if they exist in the
database.

Implementation for fetching from the database TODO.

* Implement get_user

Gets the user from the database that owns the credentials

* Fix compilation errors

* Implement find_credentials_by_card_id

* Build profile for authentication

* Implement is_access_granted

* Add audit logging on auth event

* Add state to lbgipod message handling

Allows for doorman module to have door schedules - doorman
tries to check door states during start up (if schedules are set).
FGPIO::isOn() called with "STATE" command. LIBGPIOD did not have
this message handled.

This fix now has libgpiod handle the "STATE" message and sends
the state back as "ON" or "OFF".

* Add noise filtering in authentication

Sometimes kickback will occur due to the nature of
inductors. Solenoid strikers send kicback through
the lines when they close. Additionally, wiegand
lines can be very sensitive to noise. As a result,
noise can come into the D0 and D1 lines and result
in false negatives.

This commit adds config variables to allow the user
to ignore a certain amount of bits coming in that
they might consider noise to have the software
ignore it as an access attempt.

* Update auth-db module readme

* Remove uncessary method

No longer needed since authdb tables are set up in
leosac_db creation

* Add targets to authentication mapping

No longer allows access to all doors

* Check users and groups for access

* Fix EventType values for proper FlagSet indexing

Replaced invalid bitmask hex values with sequential indicies.
The bitmask hex values were all defaulting to 0.

* Remove required AuditEntry parent

Parent is allowed to be null for authevents since it
can be a top level event.

* Add clear_schedules

Used as a helper to clear schedules so doorman can
constantly check if each door has updated schedules.

* Begin db logic for door schedules

* Fetch doors to configure schedule for

* Add open door schedules from db to doors

* Begin logic for adding close door schedules

* Add 24_7 closed schedule to doors with no schedule

* Group and insert closed full days

Groups timeframes by days, then checks if a full day
should be closed

* Sort tf vector for future creation of closed schedule

* Improve tf var name clarity

* Add closed timeframes to door schedule

* Log door open and closed schedules

* Refresh schedules every 60 seconds

* Fix closed tfs not adding correctly

* Adjust closed timeframe start/end times

Prevents closed start/end being the same as open start/end
of open schedules. Causes a 1 minute conflict without this.

* Set logging level through config

* Toggle logging through config for websockapi

* Fix bits threshold to greater/less than or equal

Config states bits_x_threshold checks less than
or equal to. While the code was only checking
less than (and vise versa).

* Fix inverse schedule logic

Was missing the first two schedule comparison

* Fix duplicate doorman objects

For each instance, an object for every door in the config was created.

This makes it so only objects for the doors under the instance are
created.

* Prevent always closed from denying access events

Before - always_closed schedules would prevent access to
the door even if the user was authorized. I've removed this
because this should be configured instead. If the door
shouldn't have access to it then this should be configured
in the schedule mapping itself for the user(s) that should
not be allowed access during that time.

Ultimately, there likely is never going to be a door
that should be denied access by all users ever. And if
there is, then it can be configured so in the schedule
mapping to prevent user access during a certain timeframe.

* Fix cmake for auth-db

* Fix return

* Update cmake

Related to commit af2fd4f

* Fix build dependency order

Link leosac_db and lib properly, and set doorman to require
them as well since doorman now works with the database

* Merge profile schedules

Old code was temporarily only returning the first
profile it found. This fixes this by merging
all schedules into one.